Move the builder files under a "builder" folder.
It was starting to make a lot of different files under the same directory, it
will be easier to follow like that.
