+++ formbuilder/actions/fieldlist.js
@@ -5,7 +5,7 @@ export const FIELD_INSERT = "FIELD_INSERT";
 export const FIELD_SWAP = "FIELD_SWAP";
 export const FORM_RESET = "FORM_RESET";
 export const FORM_UPDATE_PROPERTIES = "FORM_UPDATE_PROPERTIES";
-export const FORM_SET_SCHEMA = "FORM_SET_SCHEMA";
+
 
 export function addField(field) {
   return {type: FIELD_ADD, field};
