+++ formbuilder/containers/builder/FieldListContainer.js
@@ -1,8 +1,8 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
-import FieldList from "../components/FieldList";
-import * as FieldListActions from "../actions/fieldlist";
-import config from "../config";
+import FieldList from "../../components/builder/FieldList";
+import * as FieldListActions from "../../actions/fieldlist";
+import config from "../../config";
 
 function mapStateToProps(state) {
   return {
