+++ formbuilder/reducers/index.js
@@ -2,14 +2,14 @@ import { combineReducers } from "redux";
 
 import form from "./form";
 import notifications from "./notifications";
-import publicationStatus from "./publicationStatus";
+import serverStatus from "./serverStatus";
 import records from "./records";
 
 
 const rootReducer = combineReducers({
   notifications,
   form,
-  publicationStatus,
+  serverStatus,
   records
 });
