+++ test/components/EditableField_test.js
@@ -6,7 +6,7 @@ import { Simulate } from "react-addons-test-utils";
 
 import { createComponent } from "../test-utils";
 import config from "../../formbuilder/config";
-import EditableField from "../../formbuilder/components/EditableField";
+import EditableField from "../../formbuilder/components/builder/EditableField";
 
 
 const {fieldList} = config;
